% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R, R/kbal.R
\name{biasbound}
\alias{biasbound}
\title{Get the bound on the bias due to incomplete balance}
\usage{
biasbound(D, w, V, a, hilbertnorm = 1)

biasbound(D, w, V, a, hilbertnorm = 1)
}
\arguments{
\item{D}{xxx .}

\item{w}{numeric vector of length equal to the total number of units containing the weight for each corresponding unit.}

\item{V}{xxx .}

\item{a}{xxx .}

\item{hilbertnorm}{numeric value of the hilbertnorm.}

\item{observed}{a numeric vector of length equal to the total number of units where sampled units take a value of 1 and population units take a value of 0.}

\item{target}{a numeric vector of length equal to the total number of units where population units take a value of 1 and sample units take a value of 0.}

\item{svd.out}{The object output from \code\link{svd} performed on the kernel matrix.}

\item{w}{xxx .}

\item{hilbertnorm}{xxx .}
}
\description{
Calculate the upper bound on the bias induced by approximate balance

XXX
}
\examples{
biasbound=(observed=xxx, target=xxx, svd.out = xxx, w = xxx, hilbertnorm=1)
biasbound=(D=D, w=w, V=svd.out$u, a=svd.out$d, hilbertnorm=1)
}
