% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kbal.R
\name{biasbound_kbal}
\alias{biasbound_kbal}
\title{Get the bound on the bias due to incomplete balance}
\usage{
biasbound_kbal(D, w, V, a, hilbertnorm = 1)
}
\arguments{
\item{D}{xxx .}

\item{w}{xxx .}

\item{V}{xxx .}

\item{a}{xxx .}

\item{hilbertnorm}{xxx .}
}
\description{
XXX
}
\examples{
biasbound=(D=D, w=w, V=svd.out$u, a=svd.out$d, hilbertnorm=1)
}
