% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/kbal.R
\name{buildgauss}
\alias{buildgauss}
\title{Build Gaussian kernel matrix}
\usage{
buildgauss(X, sigma = NULL)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

