% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dimw}
\alias{dimw}
\title{Difference in Means and Difference in Weighted Means}
\usage{
dimw(X, w, target)
}
\arguments{
\item{X}{matrix of data where rows are observations and columns are covariates.}

\item{w}{numeric vector of weights for each observation.}

\item{target}{numeric vector of length equal to the total number of units where population units take a value of 1 and sample units take a value of 0.}
}
\value{
\item{dim}{the simple, unweighted difference in means.}
\item{dimw}{the weighted difference in means.}
}
\description{
Calcuates the simple difference in means or weighted difference in means between the control or sample population and the treated or target popultion.
}
\examples{
#let's say we want to get the unweighted DIM and the weighted DIM using weights from the kbal
#function with the lalonde data:
#load and clean data a bit
data(lalonde)
lalonde$nodegr=as.numeric(lalonde$educ<=11)
xvars=c("age","black","educ","hisp","married","re74","re75","nodegr","u74","u75")

#get the kbal weights
kbalout= kbal(allx=lalonde[,xvars],
               sampledinpop=FALSE,
               treatment=lalonde$nsw)
 #now use dimw to get the DIMs
 dimw(X = lalonde[,xvars], w = kbalout$w, target = lalonde$nsw)
}
