% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/el_custom.R
\name{el_custom}
\alias{el_custom}
\title{A custsom built EL optimizer.}
\usage{
el_custom(z, sumw.tol = 0.05, eps = NULL)
}
\arguments{
\item{z}{Matrix of constraints -- weighted sum must equal 0.}

\item{eps}{Used to deal with problem of taking a log near 0.}
}
\description{
Custom built EL optimizer.
}
