% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kbal.R
\name{get.dist}
\alias{get.dist}
\title{Find weights and compute L1 distance.}
\usage{
get.dist(numdims, D, Kpc, K, K_t, K_c, method, treatdrop, ...)
}
\description{
Get's the weights at the desired settings and computes
the objective function, L1.
}
