% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/kbal.R
\name{kbal}
\alias{kbal}
\title{kbal main function}
\usage{
kbal(X, D, K = NULL, whiten = TRUE, trimratio = NULL, numdims = NULL,
  maxnumdims = NULL, minnumdims = NULL, sigma = NULL)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
cat_function()
}
\keyword{cats}

