% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kbal.R
\name{kbal_meanfirst}
\alias{kbal_meanfirst}
\title{Experimental KBAL that insists on mean balance first.}
\usage{
kbal_meanfirst(X, D, K = NULL, whiten = FALSE, trimratio = NULL,
  numdims = NULL, maxnumdims = NULL, minnumdims = NULL, sigma = NULL,
  method = "ebal")
}
\arguments{
\item{X}{The original covariate data, as a numeric matrix.}

\item{D}{The treatment assignment variable taking values of 1 for treatet units and 0 for control units.}

\item{K}{Optional user-provided kernel matrix. Typically this is not user-specified, but rather is computed internally by a call to \code{buildgauss}.}

\item{whiten}{Optional pre-whitening of the data prior to construction of K. If used, rotates the data by \code{solve(chol(var(X)))}, then centers and rescales.}

\item{trimratio}{Optional \code{logical}}

\item{numdims}{Optional user-specified number of projectionss of \code{K} to balance upon.}

\item{maxnumdims}{Optional user-specified choice for the maximum number of projectsion ffo \code{K} to attempt balance on. Defaults to the number of control units.}

\item{minnumdims}{Optional user-specified choice for the minimum number of projections of \code{K}. Defualts to 2.}

\item{sigma}{Optional user-specificied paramater for the Gaussian kernel. If blank, defaults to \code{nrow(X)}.}

\item{method}{"ebal" or "el". Whether balance should be obtained on each projection of \code{K} using entropy balancing ("ebal", default) or empirical likelihood ("el")}
}
\value{
\item{w}{The weights, taking value of 1 for treated unit and the estimated weight for each control.}
\item{L1_orig}{The L1 imbalance metric on the original data}
\item{L1_kbal}{The L1 imbalance metric on the weighted data}
\item{dist.record}{The record of L1 imbalances at each number of principap components of K balanced upon}
\item{K}{The kernel matrix used}
\item{pX_D0}{The estimated density measure for the control, as measured at each X-coordinate observed (for both treated and control units)}
\item{pX_D1}{The estimated density measure for the treated, as measured at each X-coordinate observed (for both treated and control units)}
\item{sigma}{The choice of kernel bandwidth used}
}
\description{
This is an experimental version of KBAL that begins by achieving
exact mean balance on X, before adding components from K as per the usual KBAL.
The multivariate distance metrics are still based on balance in K, not the mean balance.
The mean balance must be made exact first, then it attempts to add components of K until optimized.
}
