% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{kpop}
\alias{kpop}
\title{Kernel balancing}
\usage{
kpop(allx, useasbases = NULL, b = NULL, sampled = NULL,
  sampledinpop = NULL, treatment = NULL, ebal.tol = 1e-06,
  numdims = NULL, minnumdims = NULL, maxnumdims = NULL,
  incrementby = 1, printprogess = TRUE)
}
\arguments{
\item{sampled}{a numeric vector of length equal to the total number of units where sampled units take a value of 1 and population units take a value of 0.}

\item{treatment}{a numeric vector of length equal to the total number of units where population units take a value of 1 and sample units take a value of 0.}

\item{ebal.tol}{tolerance level used by \code{ebalance}}

\item{printprogress}{}
}
\value{
\code{dist.record}

\code{biasbound}

\code{dist.sequence}

\code{numdims}

\code{biasbound.opt}
}
\description{
Th
}
