% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/lalondedata.R
\docType{data}
\name{lalonde}
\alias{lalonde}
\title{Dehejia and Wahba (1999) sample of ata from Lalonde (1986)}
\source{
\url{http://www.diamondse.info/}
}
\usage{
data(lalonde)
}
\description{
This data set includes 185 treated units
from the National Supported Work program, paired with 2490 control unites
drawn from the Panel Study of Income Dynamics (PSID-1).
}

