% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kbal.R
\name{multdiag}
\alias{multdiag}
\title{Function to multiply a square matrix, X, with a diagonal matrix, diag(d)}
\usage{
multdiag(X, d)
}
\arguments{
\item{X}{A numeric matrix of data, typically a design matrix.}

\item{d}{a vector, such that you want to know X%*%diag(d)}
}
\description{
Fast multiplication by a diagonal matrix
}
\examples{
Xd=multdiag(X,d)
}
