% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{one_hot}
\alias{one_hot}
\title{One-Hot Encoding for Categorical Data}
\usage{
one_hot(data)
}
\arguments{
\item{data}{a dataframe or matrix where columns are string or factor type covariates}
}
\value{
\item{onehot_data}{a matrix of combined sample and population data with rows corresponding to units and columns one-hot encoded categorical covariates}
\item{sampled}{a numeric vector indexing which rows off onehot_data are sample units. Of length equal to the total number of units where sampled/control units take a value of 1 and population/treated units take a value of 0.}
}
\description{
Converts raw categorical string/factor sample and population data into numeric one-hot encoded full data matrix to be passed to \code{kbal} argument \code{allx}
}
\examples{
\donttest{XX Fill in XX }
}
