% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kbal.R
\name{buildgauss}
\alias{buildgauss}
\title{Build Gaussian kernel matrix.}
\usage{
buildgauss(X, sigma = NULL)
}
\arguments{
\item{X}{A numeric matrix of data, typically a design matrix.}

\item{sigma}{The kernel ``bandwidth''. If NULL, defaults to ncol(X)}
}
\description{
Centers and rescales X then computes Guassian
kernel matrix. Entry {i,j} correspond to k(X_i,X_j) where k is the (Gaussian) kernel.
}
\examples{
K=buildgauss(rnorm(10))
}

