% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dimw}
\alias{dimw}
\title{Difference in means and Difference in weighted means}
\usage{
dimw(X, w, target)
}
\arguments{
\item{X}{Matrix of data where rows are observations and columns are covariates.}

\item{w}{numeric vector of weights for each observation.}

\item{target}{a numeric vector of length equal to the total number of units where population units take a value of 1 and sample units take a value of 0.}
}
\value{
\code{dim} the simple, unweighted difference in means

\code{dimw} the weighted difference in means
}
\description{
Calcuates the simple difference in means or weighted difference in means between the treated/sample population and the control/target popultion.
}
