% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{drop_multicollin}
\alias{drop_multicollin}
\title{Drop Multicollinear Columns}
\usage{
drop_multicollin(allx)
}
\arguments{
\item{allx}{a matrix of data to check for multicollinearity}
}
\value{
\item{allx_noMC}{resulting data matrix of full rank after multicolliear columns have been dropped}
\item{dropped_cols}{column names of those dropped}
}
\description{
Drops multicollienar columns in order of highest correlation
}
\examples{
\donttest{XX Fill in XX }
}
