% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{drop_multicollin}
\alias{drop_multicollin}
\title{Drop Multicollinear Columns}
\usage{
drop_multicollin(allx)
}
\arguments{
\item{allx}{a matrix of data to check for multicollinearity}
}
\value{
\item{allx_noMC}{resulting data matrix of full rank after multicolliear columns have been dropped}
\item{dropped_cols}{column names of those dropped}
}
\description{
Drops multicollienar columns in order of highest correlation
}
\examples{
\donttest{
# make data with multicollinearity 
data <- data.frame(x = rnorm(100),
                   y = sample.int(100,100), 
                   z = runif(100, 3,6))
test = data.frame(mc_1 = data$x,
                  mc_2 = data$x*2 + data$y - data$z)
dat = cbind(test, data)
#run
mc_check= drop_multicollin(dat)
mc_check$dropped_cols }
}
