% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ebalance.R
\name{ebalance_custom}
\alias{ebalance_custom}
\title{Modified version of ebalance (originally from Jens Hainmueller)}
\usage{
ebalance_custom(Treatment, X, base.weight = NULL, norm.constant = NULL,
  coefs = NULL, max.iterations = 200, constraint.tolerance = 0.001,
  print.level = 0)
}
\description{
Chooses weights on controls to make
covariate means equal to those of treated
}
