% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{getw}
\alias{getw}
\title{Find weights using entropy balancing.}
\usage{
getw(target, observed, allrows, ebal.tol = 1e-06, ...)
}
\arguments{
\item{target}{a numeric vector of length equal to the total number of units where population units take a value of 1 and sample units take a value of 0.}

\item{observed}{a numeric vector of length equal to the total number of units where sampled units take a value of 1 and population units take a value of 0.}

\item{allrows}{Matrix whose columns contain the left singular vectors of the kernel matrix}

\item{ebal.tol}{tolerance level used by \code{ebalance}}
}
\value{
\item{w}{numeric vector of weights}
}
\description{
Using entropy balancing to find and return the weights such that XXXX
}
