% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{makeK}
\alias{makeK}
\title{Build the Kernel Matrix}
\usage{
makeK(allx, useasbases = NULL, b = NULL)
}
\arguments{
\item{allx}{A data matrix containing all observations where rows are units and columns are covariates.}

\item{useasbases}{Binary vector argument to specify what bases to use when constructing the kernel matrix and finding weights. While the number of observations is under 2000, the default maximum is to use all observations. Due to the computation burden, when the number of observations is over 2000, the default is to use sampled units.}

\item{b}{Scaling factor in the calculation of gaussian kernel distance equivalent to the entire denominator \eqn{2\sigma^2} of the exponent. Default is twice the number of covariates or columns in \code{allx}.}
}
\value{
\item{K}{The Kernel matrix}
}
\description{
Builds the kernel matrix using Rcpp.
}
